+++ lix/packages/client/src/git/__checkout.js
@@ -31,7 +31,7 @@ import { WORKDIR } from "../../vendored/isomorphic-git/index.js"
  * @returns {Promise<void>} Resolves successfully when filesystem operations are complete
  *
  */
-export async function _checkout({
+export async function __checkout({
 	fs,
 	cache,
 	onProgress,
@@ -53,7 +53,10 @@ export async function _checkout({
 		// TODO: Figure out what to do if both 'ref' and 'remote' are specified, ref already exists,
 		// and is configured to track a different remote.
 	} catch (err) {
-		if (ref === "HEAD") throw err
+		if (ref === "HEAD") {
+			throw err
+		}
+
 		// If `ref` doesn't exist, create a new remote tracking branch
 		// Figure out the commit to checkout
 		const remoteRef = `${remote}/${ref}`