+++ lix/packages/client/src/git/add.ts
@@ -1,7 +1,9 @@
 import isoGit from "../../vendored/isomorphic-git/index.js"
-import type { RepoContext } from "../repoContext.js"
+import type { RepoContext, RepoState } from "../openRepository.js"
 
-export async function add(ctx: RepoContext, filepath: string | string[]) {
+export async function add(ctx: RepoContext, state: RepoState, filepath: string | string[]) {
+	await state.ensureFirstBatch()
+
 	return await isoGit.add({
 		fs: ctx.rawFs,
 		parallel: true,