+++ lix/packages/client/src/lix/checkoutPlaceholders.ts
@@ -1,14 +1,24 @@
 import type { RepoState, RepoContext } from "../openRepository.js"
 import isoGit from "../../vendored/isomorphic-git/index.js"
-import { doCheckout } from "../git/checkout.js"
+import { _checkout } from "../git/checkout.js"
 import { modeToFileType } from "../git/helpers.js"
 
-export async function checkOutPlaceholders(ctx: RepoContext, state: RepoState) {
+export async function checkOutPlaceholders(
+	ctx: RepoContext,
+	state: RepoState,
+	{
+		materializeGitignores = true,
+		preload = [],
+	}: {
+		materializeGitignores?: boolean
+		preload?: string[]
+	} = {}
+) {
 	const { rawFs, cache, dir } = ctx
 	const { branchName, checkedOut, sparseFilter } = state
 
-	await doCheckout({
-		fs: rawFs,
+	await _checkout({
+		fs: ctx.rawFs,
 		cache,
 		dir,
 		ref: branchName,
@@ -18,6 +28,7 @@ export async function checkOutPlaceholders(ctx: RepoContext, state: RepoState) {
 	const fs = rawFs
 	const gitignoreFiles: string[] = []
 
+	let rootHash: string | undefined
 	await isoGit.walk({
 		fs,
 		dir,
@@ -29,14 +40,19 @@ export async function checkOutPlaceholders(ctx: RepoContext, state: RepoState) {
 			if (!commit) {
 				return undefined
 			}
-			const fileMode = await commit.mode()
-
-			const fileType = modeToFileType(fileMode)
 
 			if (fullpath.endsWith(".gitignore")) {
 				gitignoreFiles.push(fullpath)
-				return undefined
 			}
+
+			const fileMode = await commit.mode()
+			const oid = await commit.oid()
+
+			if (fullpath === ".") {
+				rootHash = oid
+			}
+
+			const fileType = modeToFileType(fileMode)
 
 			if (
 				sparseFilter &&
@@ -53,12 +69,12 @@ export async function checkOutPlaceholders(ctx: RepoContext, state: RepoState) {
 			}
 
 			if (fileType === "file" && !checkedOut.has(fullpath)) {
-				await fs._createPlaceholder(fullpath, { mode: fileMode })
+				await fs._createPlaceholder(fullpath, { mode: fileMode, oid, rootHash })
 				return fullpath
 			}
 
 			if (fileType === "symlink" && !checkedOut.has(fullpath)) {
-				await fs._createPlaceholder(fullpath, { mode: fileMode })
+				await fs._createPlaceholder(fullpath, { mode: fileMode, oid, rootHash })
 				return fullpath
 			}
 
@@ -67,16 +83,12 @@ export async function checkOutPlaceholders(ctx: RepoContext, state: RepoState) {
 		},
 	})
 
-	if (gitignoreFiles.length) {
-		await doCheckout({
-			fs: rawFs,
-			dir,
-			cache,
-			ref: branchName,
-			filepaths: gitignoreFiles,
-		})
-		gitignoreFiles.map((file) => checkedOut.add(file))
+	// materializeGitignores is only false for testing when opeinging a snapshot and using emptyWorkdir, using lazyFs will hang forewer but we allready have the ignore files in object store
+	if (gitignoreFiles.length && materializeGitignores) {
+		preload = [...gitignoreFiles, ...preload]
 	}
 
-	state.pending && (await state.pending)
+	await state.ensureFirstBatch({ preload })
+
+	return { gitignoreFiles }
 }