+++ inlang/development-projects/inlang-nextjs/CHANGELOG.md
@@ -1,5 +1,38 @@
 # next-js-testapp
 
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [2806698]
+  - @inlang/cli@2.18.0
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [3896fe6]
+  - @inlang/cli@2.17.0
+
+## 0.1.18
+
+### Patch Changes
+
+- @inlang/cli@2.16.10
+
+## 0.1.17
+
+### Patch Changes
+
+- @inlang/cli@2.16.9
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [a27b7a4]
+  - @inlang/cli@2.16.8
+
 ## 0.1.15
 
 ### Patch Changes