+++ inlang/source-code/cli/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @inlang/cli
 
+## 2.18.0
+
+### Minor Changes
+
+- 2806698: force release cli, as it was missing from the changeset targets for some reason
+
+## 2.17.0
+
+### Minor Changes
+
+- 3896fe6: fix: move dependencies to dev dependencies to avoid "unpublished dependency" issues
+
+## 2.16.10
+
+### Patch Changes
+
+- Updated dependencies [00ad046]
+- Updated dependencies [00ad046]
+  - @lix-js/fs@2.0.0
+  - @lix-js/client@2.0.0
+  - @inlang/sdk@0.35.1
+
+## 2.16.9
+
+### Patch Changes
+
+- Updated dependencies [ae47203]
+  - @inlang/sdk@0.35.0
+
+## 2.16.8
+
+### Patch Changes
+
+- a27b7a4: This reintroduces reactivity to lint reports - see https://github.com/opral/monorepo/pull/2792 for more details
+- Updated dependencies [d27a983]
+- Updated dependencies [a27b7a4]
+  - @inlang/sdk@0.34.10
+
 ## 2.16.7
 
 ### Patch Changes