+++ inlang/source-code/paraglide/paraglide-astro/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @inlang/paraglide-astro
 
+## 0.1.22
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.62
+
+## 0.1.21
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.61
+
+## 0.1.20
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.60
+
+## 0.1.19
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.59
+
+## 0.1.18
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.58
+
+## 0.1.17
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.57
+
+## 0.1.16
+
+### Patch Changes
+
+- @inlang/paraglide-vite@1.2.56
+
 ## 0.1.15
 
 ### Patch Changes