+++ inlang/source-code/plugins/json/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @inlang/plugin-json
 
+## 5.1.46
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 5.1.45
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 5.1.44
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+
+## 5.1.43
+
+### Patch Changes
+
+- @inlang/sdk@0.35.1
+
+## 5.1.42
+
+### Patch Changes
+
+- Updated dependencies [ae47203]
+  - @inlang/sdk@0.35.0
+
+## 5.1.41
+
+### Patch Changes
+
+- Updated dependencies [d27a983]
+- Updated dependencies [a27b7a4]
+  - @inlang/sdk@0.34.10
+
 ## 5.1.40
 
 ### Patch Changes