+++ inlang/source-code/rpc/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @inlang/rpc
 
+## 0.3.26
+
+### Patch Changes
+
+- @inlang/sdk@0.35.4
+
+## 0.3.25
+
+### Patch Changes
+
+- @inlang/sdk@0.35.3
+
+## 0.3.24
+
+### Patch Changes
+
+- @inlang/sdk@0.35.2
+
+## 0.3.23
+
+### Patch Changes
+
+- @inlang/sdk@0.35.1
+
+## 0.3.22
+
+### Patch Changes
+
+- Updated dependencies [ae47203]
+  - @inlang/sdk@0.35.0
+
+## 0.3.21
+
+### Patch Changes
+
+- Updated dependencies [d27a983]
+- Updated dependencies [a27b7a4]
+  - @inlang/sdk@0.34.10
+
 ## 0.3.20
 
 ### Patch Changes